xfs_scrub: create online filesystem scrub program
Create a filesystem scrubbing tool that walks the directory tree,
queries every file's extents, extended attributes, and stat data.  For
generic (non-XFS) filesystems this depends on the kernel to do nearly
all the validation.  Optionally, we can (try to) read all the file
data.

For XFS, we perform sequential scans of each AG's metadata, inodes,
extent maps, and file data.  Being XFS specific, we can work with
the in-kernel scrubbers to perform much stronger
metadata checking and cross-referencing.  We can also take advantage
of newer ioctls such as GETFSMAP to perform faster read verification.

In the future we will be able to take advantage of (still unwritten)
features such as parent directory pointers to fully validate all
metadata.  However, this tool /should/ work for most non-XFS
filesystems such as ext4 and btrfs.

Note also that the scrub tool can shut down the filesystem if errors
are found.  This is not a default option since scrubbing is very
immature at this time.  It can also ask the XFS driver in the kernel
to optimize or repair metadata, though this may not be successful.

Signed-off-by: Darrick J. Wong <darrick.wong@oracle.com>
---
v2: Add command line options so that we can abort after a certain
number of errors; or shut down the fs if errors are found.
v3: Add the ability to shut down on error, optimize (preen) some
metadata, and repair some metadata too.